Ensure correctness with Math.toIntExact()

+++ src/main/java/org/sharetrace/experiment/Experiment.java
@@ -283,7 +283,7 @@ public abstract class Experiment implements Runnable {
   }
 
   protected int cacheIntervals() {
-    return (int) (2 * defaultTtl().toDays());
+    return Math.toIntExact(2 * defaultTtl().toDays());
   }
 
   protected Duration cacheRefreshPeriod() {
